--- .env.example
@@ -1,23 +1,0 @@
-# Example environment variables for QA tests
-# Copy this file to `.env` and fill in secret values. Do NOT commit your local `.env`.
-
-STUDENT_EMAIL=
-STUDENT_PASSWORD=
-MENTOR_EMAIL=
-MENTOR_PASSWORD=
-ADMIN_EMAIL=
-ADMIN_PASSWORD=
-BASE_URL=https://stage.tripleten.percruit.com/
-
-
-# UI / Playwright
-UI_BASE_URL=http://localhost:3000
-UI_USER_EMAIL=example@example.com
-UI_USER_PASSWORD=changeme
-PLAYWRIGHT_HEADLESS=true
-
-# API / Pytest
-API_BASE_URL=http://localhost:8000
-API_USER_EMAIL=example@example.com
-API_USER_PASSWORD=changeme
-API_TOKEN=
