+++ ui_tests/playwright.config.ts
@@ -9,8 +9,8 @@
 //   reporter: [['html', { open: 'never' }]],
 // });
 
-import { chromium, defineConfig, devices } from "@playwright/test";
-import dotenv from "dotenv";
+import { chromium, defineConfig, devices } from '@playwright/test';
+import dotenv from 'dotenv';
 /**
  * Read environment variables from file.
  * https://github.com/motdotla/dotenv
@@ -21,7 +21,7 @@ dotenv.config();
  * See https://playwright.dev/docs/test-configuration.
  */
 export default defineConfig({
-  testDir: "./tests",
+  testDir: './tests',
   /* Run tests in files in parallel */
   fullyParallel: true,
   /* Fail the build on CI if you accidentally left test.only in the source code. */
@@ -36,41 +36,41 @@ export default defineConfig({
   /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
   use: {
     /* Base URL to use in actions like `await page.goto('/')`. */
-    baseURL: "https://stage.tripleten.percruit.com",
+    baseURL: 'https://stage.tripleten.percruit.com',
 
     /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
-    trace: "on-all-retries",
+    trace: 'on-all-retries',
 
-    headless: true,
+    headless: false,
 
-    testIdAttribute: "data-tt-tour",
+    testIdAttribute: 'data-tt-tour',
 
-    video: "off",
+    video: 'off',
 
-    screenshot: "on",
+    screenshot: 'on',
   },
 
   /* Configure projects for major browsers */
   projects: [
     {
-      name: "chromium",
+      name: 'chromium',
       use: {
-        ...devices["Desktop Chrome"]
-      }
+        ...devices['Desktop Chrome'],
+      },
     },
 
     {
-      name: "firefox",
-       use: {
-        ...devices["Desktop Firefox"]
-      }
+      name: 'firefox',
+      use: {
+        ...devices['Desktop Firefox'],
+      },
     },
 
     {
-      name: "webkit",
-       use: {
-        ...devices["Desktop Safari"]
-      }
+      name: 'webkit',
+      use: {
+        ...devices['Desktop Safari'],
+      },
     },
   ],
 });
